requirements.txt
python-telegram-bot
flask
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
gunicorn
